backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
torch==2.4.1
torchvision==0.19.1
langchain==0.3.1
langchain-groq==0.3.8
langchain-huggingface==0.1.0
langchain-community==0.3.1
supabase==2.12.0
pypdf==5.1.0
python-docx==1.1.2
pandas==2.2.3
requests==2.32.3
python-multipart==0.0.9
soundfile==0.12.1
coqui-tts==0.22.0
python-dotenv==1.0.1
groq==0.11.0
pillow==10.4.0